Fixed start tags by removing spaces
It seems that like IntelliJ code-fold, putting space at beginning of start tag, to make text align w/ that of end(" " for start "/" for end), causes A problem. Here it makes start tag visible. In intelliJ it makes fold-start invalid, so I just put space before "<" on start, so that it's 1 to the right of it's counterpart on end, but all text to the right is aligned. Here(in at least Chrome's rendering engine) I guess I haven't found A way to make code mor aesthetic/readable in that way, on that type of thing(start tags).